change int to int16 and perform data alignment

+++ internal/domain/dto/fuzzyArtifact.go
@@ -1,8 +1,8 @@
 package dto
 
 type FuzzyArtifactInput struct {
-	IntervalStart int `json:"intervalStart"`
-	IntervalEnd   int `json:"intervalEnd"`
+	IntervalStart int16 `json:"intervalStart"`
+	IntervalEnd   int16 `json:"intervalEnd"`
 }
 
 type FuzzyArtefactOutput struct {
